Allow to use boolean values as min and max values.

+++ test/integer_infinite_numberpicker_test.dart
@@ -91,8 +91,8 @@ void main() {
 
 Future<NumberPicker> _testNumberPicker(
     {WidgetTester tester,
-    int minValue,
-    int maxValue,
+    double minValue,
+    double maxValue,
     int initialValue,
     int scrollBy,
     int step = 1,
